Remove attest dependency
`attest` as referenced here is attempting to pull from the npm registry, which only has an attest version up to 1.0.1. And so anything attempting to install this component as a dependency will fail during `npm i`.

It appears this was added here as a mistake, so this PR removes it.